--- .eslintignore
@@ -1,24 +1,0 @@
-__fixtures__
-__mocks__
-dist
-node_modules
-.yarn
-.history
-build
-coverage
-examples/
-
-packages/lqip-loader/lib/
-packages/docusaurus/lib/
-packages/docusaurus-*/lib/*
-packages/eslint-plugin/lib/
-packages/stylelint-copyright/lib/
-
-packages/create-docusaurus/lib/*
-packages/create-docusaurus/templates/facebook
-
-website/_dogfooding/_swizzle_theme_tests
-website/_dogfooding/_asset-tests/badSyntax.js
-
-
-packages/docusaurus-plugin-ideal-image/src/theme/IdealImageLegacy
