+++ package.json
@@ -79,6 +79,7 @@
   "devDependencies": {
     "@ai-sdk/react": "^2.0.30",
     "@crowdin/cli": "^4.14.2",
+    "@eslint/js": "^10.0.1",
     "@swc/core": "^1.15.32",
     "@testing-library/dom": "^10.4.1",
     "@testing-library/jest-dom": "^6.9.1",
@@ -90,34 +91,34 @@
     "@types/react": "^19.2.14",
     "@types/semver": "^7.7.1",
     "@types/shelljs": "^0.8.12",
-    "@typescript-eslint/eslint-plugin": "^5.62.0",
-    "@typescript-eslint/parser": "^5.62.0",
+    "@vitejs/plugin-react": "^5.0.0",
+    "@vitest/eslint-plugin": "^1.6.17",
     "cross-env": "^10.1.0",
     "cspell": "^8.18.1",
-    "eslint": "^8.45.0",
-    "eslint-config-airbnb": "^19.0.4",
-    "eslint-config-prettier": "^8.8.0",
+    "eslint": "^10.3.0",
+    "eslint-config-prettier": "^10.1.8",
     "eslint-plugin-header": "^3.1.1",
-    "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-jsx-a11y": "^6.7.1",
-    "eslint-plugin-react": "^7.32.2",
-    "eslint-plugin-react-compiler": "^19.0.0-beta-40c6c23-20250301",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-regexp": "^1.15.0",
+    "eslint-plugin-import": "^2.32.0",
+    "eslint-plugin-jsx-a11y": "^6.10.2",
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-react-hooks": "^7.1.1",
+    "eslint-plugin-regexp": "^3.1.0",
+    "globals": "^17.6.0",
     "husky": "^9.1.7",
     "image-size": "^2.0.2",
     "jest-serializer-ansi-escapes": "^5.0.0",
     "jest-serializer-react-helmet-async": "^1.0.21",
+    "jiti": "^2.7.0",
     "jsdom": "^25.0.1",
     "lerna": "^7.4.2",
     "lerna-changelog": "^2.2.0",
     "lint-staged": "^17.0.2",
     "lockfile-lint": "^5.0.0",
     "npm-run-all": "^4.1.5",
+    "oxfmt": "^0.47.0",
     "patch-package": "^8.0.1",
     "pkg-pr-new": "^0.0.68",
     "postinstall-postinstall": "^2.1.0",
-    "oxfmt": "^0.47.0",
     "react": "^19.2.5",
     "react-dom": "^19.2.5",
     "rimraf": "^3.0.2",
@@ -128,9 +129,8 @@
     "stylelint-config-standard": "^29.0.0",
     "syncpack": "^14.3.1",
     "typescript": "~6.0.3",
-    "vitest": "^4.0.0",
-    "@vitejs/plugin-react": "^5.0.0",
-    "@vitest/eslint-plugin": "^1.4.0"
+    "typescript-eslint": "^8.59.3",
+    "vitest": "^4.0.0"
   },
   "resolutions": {
     "**/pretty-format/react-is": "^19.2.0",
